Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
 
+
 int main() {
-    int a,b,result;
+    float a,b,result;
     char ch;
     scanf("%d %d %c",&a, &b, &ch);
     switch(ch) {
